Get table of absolute percent cover for given park and year

Get table of absolute percent cover for given park and year

get_pct_cover_tbl(cfg, park, year)

Arguments

cfg
NPS configuration list object; see get_nps_config

park
park abbreviation, eg "CABR", "CHIS" or "SAMO"

year
4-digit year

Value

Tibble that reproduces from mod_ExportQueries:

Export_AnnualReport_AbsoluteCover()

for "Figure E.2. Absolute foliar cover (%) of plant growth forms, as observed during 20XX monitoring at CABR. Colored bars show mean values, while error bars extend ±1 s.d. from the means." from MEDN_veg_protocol_NARRATIVE_FINAL_8Sep2016.pdf.

Examples

cfg  <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
park <- "CABR"
year <- 2015

get_pct_cover_tbl(cfg, park, year)

#> Parsed with column specification:
#> cols(
#>   AnnualPerennial_code = col_character(),
#>   AnnualPerennial_desc = col_character(),
#>   Sort_order = col_integer(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Nativity_code = col_character(),
#>   Nativity_desc = col_character(),
#>   Sort_order = col_integer(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_integer(),
#>   Location_ID = col_integer(),
#>   Event_Group_ID = col_integer(),
#>   Analysis_code = col_integer(),
#>   Is_excluded = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Taxon_ID = col_integer(),
#>   Sort_order = col_double(),
#>   TSN = col_double(),
#>   LocalAcceptedTSN = col_integer(),
#>   Taxon_is_active = col_integer(),
#>   Refers_to = col_integer(),
#>   IsEndemic = col_integer(),
#>   ParkStatusID = col_integer(),
#>   AbundanceID = col_integer(),
#>   NativityID = col_integer(),
#>   InNPSpecies = col_integer(),
#>   Park_List = col_integer(),
#>   Common_picklist = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Layer_code = col_character(),
#>   Layer_desc = col_character(),
#>   Sort_order = col_integer()
#> )

#> Parsed with column specification:
#> cols(
#>   Site_ID = col_integer(),
#>   GIS_Location_ID = col_character(),
#>   Site_Name = col_character(),
#>   Site_Desc = col_character(),
#>   Unit_Code = col_character(),
#>   Site_Notes = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Location_ID = col_integer(),
#>   Site_ID = col_integer(),
#>   X_Coord = col_double(),
#>   Y_Coord = col_double(),
#>   UTM_Zone = col_integer(),
#>   No_Of_Transect_Points = col_integer(),
#>   Slope_degrees = col_integer(),
#>   Aspect_bearing = col_integer(),
#>   Elevation_m = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_integer(),
#>   Location_ID = col_integer(),
#>   Event_Group_ID = col_integer(),
#>   Analysis_code = col_integer(),
#>   Is_excluded = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Event_Point_ID = col_integer(),
#>   Event_ID = col_integer(),
#>   Protocol_ID = col_integer(),
#>   Length_cm = col_integer(),
#>   Point_No = col_integer(),
#>   Substrate = col_character(),
#>   MaxHeight_cm = col_integer(),
#>   MaxHeight_cm_DEAD = col_character(),
#>   Transect_Event_Comments = col_character(),
#>   Entered_by = col_character(),
#>   Entered_date = col_character(),
#>   Updated_by = col_character(),
#>   Updated_date = col_character(),
#>   Import_date = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   ID = col_integer(),
#>   Event_Point_ID = col_integer(),
#>   Species_Rank = col_integer(),
#>   Species_Code = col_character(),
#>   Condition = col_character(),
#>   Phenology = col_character(),
#>   Entered_by = col_character(),
#>   Entered_date = col_character(),
#>   Updated_by = col_character(),
#>   Updated_date = col_character(),
#>   Import_date = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Condition = col_character(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Site_ID = col_integer(),
#>   GIS_Location_ID = col_character(),
#>   Site_Name = col_character(),
#>   Site_Desc = col_character(),
#>   Unit_Code = col_character(),
#>   Site_Notes = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Location_ID = col_integer(),
#>   Site_ID = col_integer(),
#>   X_Coord = col_double(),
#>   Y_Coord = col_double(),
#>   UTM_Zone = col_integer(),
#>   No_Of_Transect_Points = col_integer(),
#>   Slope_degrees = col_integer(),
#>   Aspect_bearing = col_integer(),
#>   Elevation_m = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_integer(),
#>   Location_ID = col_integer(),
#>   Event_Group_ID = col_integer(),
#>   Analysis_code = col_integer(),
#>   Is_excluded = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Event_Point_ID = col_integer(),
#>   Event_ID = col_integer(),
#>   Protocol_ID = col_integer(),
#>   Length_cm = col_integer(),
#>   Point_No = col_integer(),
#>   Substrate = col_character(),
#>   MaxHeight_cm = col_integer(),
#>   MaxHeight_cm_DEAD = col_character(),
#>   Transect_Event_Comments = col_character(),
#>   Entered_by = col_character(),
#>   Entered_date = col_character(),
#>   Updated_by = col_character(),
#>   Updated_date = col_character(),
#>   Import_date = col_character()
#> )

#> # A tibble: 58 x 11
#> # Groups:   SurveyYear, Park, Vegetation_Community, FxnGroup [50]
#>    SurveyYear Park  Vegetation_Community FxnGroup Nativity NofTransects Average StdDev MinRange
#>         <int> <chr> <chr>                <chr>    <chr>           <int>   <dbl>  <dbl>    <dbl>
#>  1       2015 CABR  Black Sage Scrub     Grass    Native              9   0.556  0.882       0.
#>  2       2015 CABR  Black Sage Scrub     Grass    Nonnati…            9   0.222  0.441       0.
#>  3       2015 CABR  Black Sage Scrub     Herbace… Native              9   1.67   1.58        0.
#>  4       2015 CABR  Black Sage Scrub     Shrub    Native              9   5.11   1.27        3.
#>  5       2015 CABR  Black Sage Scrub     Sub-shr… Native              9   0.333  0.500       0.
#>  6       2015 CABR  Black Sage Scrub     Vine     Native              9   0.556  0.527       0.
#>  7       2015 CABR  Black Sage Scrub     <NA>     <NA>                9   1.22   0.441       1.
#>  8       2015 CABR  California Brittle … Grass    Native             10   0.500  0.707       0.
#>  9       2015 CABR  California Brittle … Grass    Nonnati…           10   0.400  0.516       0.
#> 10       2015 CABR  California Brittle … Herbace… Native             10   1.60   1.26        0.
#> # ... with 48 more rows, and 2 more variables: MaxRange <dbl>, Query_type <chr>