Get table of total event points for given park

Get table of total event points for given park

get_total_eventpoints_tbl(cfg, park)

Arguments

cfg
NPS configuration list object; see get_nps_config

park
park abbreviation, eg "CABR", "CHIS" or "SAMO"

Value

tibble with fields: Park, IslandCode, Location_ID, SiteCode, Vegetation_Community, SurveyYear, SurveyDate, NofPoints

Examples

cfg <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
get_total_eventpoints_tbl(cfg, park)

#> Parsed with column specification:
#> cols(
#>   Site_ID = col_integer(),
#>   GIS_Location_ID = col_character(),
#>   Site_Name = col_character(),
#>   Site_Desc = col_character(),
#>   Unit_Code = col_character(),
#>   Site_Notes = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Location_ID = col_integer(),
#>   Site_ID = col_integer(),
#>   X_Coord = col_double(),
#>   Y_Coord = col_double(),
#>   UTM_Zone = col_integer(),
#>   No_Of_Transect_Points = col_integer(),
#>   Slope_degrees = col_integer(),
#>   Aspect_bearing = col_integer(),
#>   Elevation_m = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_integer(),
#>   Location_ID = col_integer(),
#>   Event_Group_ID = col_integer(),
#>   Analysis_code = col_integer(),
#>   Is_excluded = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Event_Point_ID = col_integer(),
#>   Event_ID = col_integer(),
#>   Protocol_ID = col_integer(),
#>   Length_cm = col_integer(),
#>   Point_No = col_integer(),
#>   Substrate = col_character(),
#>   MaxHeight_cm = col_integer(),
#>   MaxHeight_cm_DEAD = col_character(),
#>   Transect_Event_Comments = col_character(),
#>   Entered_by = col_character(),
#>   Entered_date = col_character(),
#>   Updated_by = col_character(),
#>   Updated_date = col_character(),
#>   Import_date = col_character()
#> )

#> # A tibble: 173 x 8
#> # Groups:   Park, IslandCode, Location_ID, SiteCode, Vegetation_Community,
#> #   SurveyYear [?]
#>    Park  IslandCode Location_ID SiteCode Vegetation_Comm… SurveyYear SurveyDate
#>    <chr> <chr>            <int> <chr>    <chr>                 <int> <chr>     
#>  1 CABR  CABR       -2141031974 034      Black Sage Scrub       2014 4/9/2014 …
#>  2 CABR  CABR       -2141031974 034      Black Sage Scrub       2015 4/28/2015…
#>  3 CABR  CABR       -2084019541 011      California Brit…       2012 5/9/2012 …
#>  4 CABR  CABR       -2084019541 011      California Brit…       2013 4/26/2013…
#>  5 CABR  CABR       -2032028138 086      Black Sage Scrub       2017 4/14/2017…
#>  6 CABR  CABR       -2002950960 076      Chamise Coastal…       2013 4/25/2013…
#>  7 CABR  CABR       -2002950960 076      Chamise Coastal…       2014 4/8/2014 …
#>  8 CABR  CABR       -1895349370 038      California Brit…       2015 5/1/2015 …
#>  9 CABR  CABR       -1895349370 038      California Brit…       2016 4/20/2016…
#> 10 CABR  CABR       -1780577253 027      Black Sage Scrub       2014 4/10/2014…
#> # ... with 163 more rows, and 1 more variable: NofPoints <int>