Get table of absolute percent cover for given park and year

Get table of absolute percent cover for given park and year

get_pct_cover_tbl(cfg, park, year)

Arguments

cfg
NPS configuration list object; see get_nps_config

park
park abbreviation, eg "CABR", "CHIS" or "SAMO"

year
4-digit year

Value

Tibble that reproduces from mod_ExportQueries:

Export_AnnualReport_AbsoluteCover()

for "Figure E.2. Absolute foliar cover (%) of plant growth forms, as observed during 20XX monitoring at CABR. Colored bars show mean values, while error bars extend ±1 s.d. from the means." from MEDN_veg_protocol_NARRATIVE_FINAL_8Sep2016.pdf.

Examples

cfg  <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
park <- "CABR"
year <- 2015

get_pct_cover_tbl(cfg, park, year)

#> Parsed with column specification:
#> cols(
#>   Site_ID = col_double(),
#>   GIS_Location_ID = col_logical(),
#>   Site_Name = col_character(),
#>   Site_Desc = col_character(),
#>   Unit_Code = col_character(),
#>   Site_Notes = col_logical()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_logical(),
#>   Location_ID = col_double(),
#>   Location_Code = col_character(),
#>   Site_ID = col_double(),
#>   X_Coord = col_double(),
#>   Y_Coord = col_double(),
#>   Coord_Units = col_character(),
#>   Coord_System = col_character(),
#>   UTM_Zone = col_double(),
#>   Datum = col_character(),
#>   Unit_Code = col_character(),
#>   Loc_Type = col_character(),
#>   Loc_Notes = col_character(),
#>   Monitoring_Status = col_character(),
#>   Vegetation_Community = col_character(),
#>   No_Of_Transect_Points = col_double(),
#>   Slope_degrees = col_double(),
#>   Aspect_bearing = col_double(),
#>   Elevation_m = col_double(),
#>   Slope_shape = col_character(),
#>   Entered_date = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_double(),
#>   Location_ID = col_double(),
#>   Event_Group_ID = col_double(),
#>   Start_Time = col_logical(),
#>   End_Time = col_logical(),
#>   Visit_number = col_logical(),
#>   Analysis_code = col_double(),
#>   Is_excluded = col_double(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Certified_by = col_logical(),
#>   Certified_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Event_Point_ID = col_double(),
#>   Event_ID = col_double(),
#>   Protocol_ID = col_double(),
#>   Length_cm = col_double(),
#>   Point_No = col_double(),
#>   Substrate = col_character(),
#>   MaxHeight_cm = col_double(),
#>   MaxHeight_cm_DEAD = col_logical(),
#>   Transect_Event_Comments = col_logical(),
#>   Entered_by = col_logical(),
#>   Entered_date = col_character(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> Warning: 1127 parsing failures.
#>  row               col           expected actual                                                                                   file
#> 4301 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     27 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4307 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     27 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4313 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     30 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4340 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     16 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4706 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     48 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> .... ................. .................. ...... ......................................................................................
#> See problems(...) for more details.

#> Parsed with column specification:
#> cols(
#>   ID = col_double(),
#>   Event_Point_ID = col_double(),
#>   Species_Rank = col_double(),
#>   Species_Code = col_character(),
#>   Condition = col_character(),
#>   Phenology = col_logical(),
#>   Entered_by = col_logical(),
#>   Entered_date = col_character(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> Parsed with column specification:
#> cols(
#>   Condition = col_character(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Site_ID = col_double(),
#>   GIS_Location_ID = col_logical(),
#>   Site_Name = col_character(),
#>   Site_Desc = col_character(),
#>   Unit_Code = col_character(),
#>   Site_Notes = col_logical()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_logical(),
#>   Location_ID = col_double(),
#>   Location_Code = col_character(),
#>   Site_ID = col_double(),
#>   X_Coord = col_double(),
#>   Y_Coord = col_double(),
#>   Coord_Units = col_character(),
#>   Coord_System = col_character(),
#>   UTM_Zone = col_double(),
#>   Datum = col_character(),
#>   Unit_Code = col_character(),
#>   Loc_Type = col_character(),
#>   Loc_Notes = col_character(),
#>   Monitoring_Status = col_character(),
#>   Vegetation_Community = col_character(),
#>   No_Of_Transect_Points = col_double(),
#>   Slope_degrees = col_double(),
#>   Aspect_bearing = col_double(),
#>   Elevation_m = col_double(),
#>   Slope_shape = col_character(),
#>   Entered_date = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_double(),
#>   Location_ID = col_double(),
#>   Event_Group_ID = col_double(),
#>   Start_Time = col_logical(),
#>   End_Time = col_logical(),
#>   Visit_number = col_logical(),
#>   Analysis_code = col_double(),
#>   Is_excluded = col_double(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Certified_by = col_logical(),
#>   Certified_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Event_Point_ID = col_double(),
#>   Event_ID = col_double(),
#>   Protocol_ID = col_double(),
#>   Length_cm = col_double(),
#>   Point_No = col_double(),
#>   Substrate = col_character(),
#>   MaxHeight_cm = col_double(),
#>   MaxHeight_cm_DEAD = col_logical(),
#>   Transect_Event_Comments = col_logical(),
#>   Entered_by = col_logical(),
#>   Entered_date = col_character(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> Warning: 1127 parsing failures.
#>  row               col           expected actual                                                                                   file
#> 4301 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     27 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4307 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     27 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4313 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     30 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4340 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     16 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> 4706 MaxHeight_cm_DEAD 1/0/T/F/TRUE/FALSE     48 '~/Gdrive Ecoquants/projects/nps-ecoquants/data/R_tables_csv/CABR/tbl_Event_Point.csv'
#> .... ................. .................. ...... ......................................................................................
#> See problems(...) for more details.

#> Parsed with column specification:
#> cols(
#>   AnnualPerennial_code = col_character(),
#>   AnnualPerennial_desc = col_character(),
#>   Sort_order = col_double(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Nativity_code = col_character(),
#>   Nativity_desc = col_character(),
#>   Sort_order = col_double(),
#>   Analysis_code = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Event_ID = col_double(),
#>   Location_ID = col_double(),
#>   Event_Group_ID = col_double(),
#>   Start_Time = col_logical(),
#>   End_Time = col_logical(),
#>   Visit_number = col_logical(),
#>   Analysis_code = col_double(),
#>   Is_excluded = col_double(),
#>   Updated_by = col_logical(),
#>   Updated_date = col_logical(),
#>   Certified_by = col_logical(),
#>   Certified_date = col_logical(),
#>   Import_date = col_logical()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Taxon_ID = col_double(),
#>   Family_order = col_logical(),
#>   Sort_order = col_double(),
#>   List_order = col_logical(),
#>   TSN = col_double(),
#>   LocalAcceptedTSN = col_double(),
#>   Taxon_is_active = col_double(),
#>   Refers_to = col_double(),
#>   IsEndemic = col_double(),
#>   ParkStatusID = col_double(),
#>   AbundanceID = col_double(),
#>   NativityID = col_double(),
#>   InNPSpecies = col_double(),
#>   Park_List = col_double(),
#>   Taxon_code = col_logical(),
#>   Common_picklist = col_double()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   Layer_code = col_character(),
#>   Layer_desc = col_character(),
#>   Sort_order = col_double()
#> )

#> # A tibble: 58 x 11
#> # Groups:   SurveyYear, Park, Vegetation_Community, FxnGroup [50]
#>    SurveyYear Park  Vegetation_Comm… FxnGroup Nativity NofTransects Average
#>         <int> <chr> <chr>            <chr>    <chr>           <int>   <dbl>
#>  1       2015 CABR  Black Sage Scrub NA       NA                  9   1.22 
#>  2       2015 CABR  Black Sage Scrub Grass    Native              9   0.556
#>  3       2015 CABR  Black Sage Scrub Grass    Nonnati…            9   0.222
#>  4       2015 CABR  Black Sage Scrub Herbace… Native              9   1.67 
#>  5       2015 CABR  Black Sage Scrub Shrub    Native              9   5.11 
#>  6       2015 CABR  Black Sage Scrub Sub-shr… Native              9   0.333
#>  7       2015 CABR  Black Sage Scrub Vine     Native              9   0.556
#>  8       2015 CABR  California Brit… NA       NA                 10   1.1  
#>  9       2015 CABR  California Brit… Grass    Native             10   0.5  
#> 10       2015 CABR  California Brit… Grass    Nonnati…           10   0.4  
#> # … with 48 more rows, and 4 more variables: StdDev <dbl>, MinRange <dbl>,
#> #   MaxRange <dbl>, Query_type <chr>